+++ src/main/java/org/jsoup/parser/Parser.java
@@ -51,7 +51,7 @@ public class Parser {
     }
     
     public Document parseInput(String html, String baseUri) {
-        return treeBuilder.parse(new StringReader(html), baseUri, this);
+        return parseInput(new StringReader(html), baseUri);
     }
 
     public Document parseInput(Reader inputHtml, String baseUri) {
@@ -59,8 +59,13 @@ public class Parser {
     }
 
     public List<Node> parseFragmentInput(String fragment, @Nullable Element context, String baseUri) {
+        return parseFragmentInput(new StringReader(fragment), context, baseUri);
+    }
+
+    public List<Node> parseFragmentInput(Reader fragment, @Nullable Element context, String baseUri) {
         return treeBuilder.parseFragment(fragment, context, baseUri, this);
     }
+
     // gets & sets
     /**
      * Get the TreeBuilder currently in use.
@@ -184,7 +189,7 @@ public class Parser {
      */
     public static List<Node> parseFragment(String fragmentHtml, Element context, String baseUri) {
         HtmlTreeBuilder treeBuilder = new HtmlTreeBuilder();
-        return treeBuilder.parseFragment(fragmentHtml, context, baseUri, new Parser(treeBuilder));
+        return treeBuilder.parseFragment(new StringReader(fragmentHtml), context, baseUri, new Parser(treeBuilder));
     }
 
     /**
@@ -202,7 +207,7 @@ public class Parser {
         HtmlTreeBuilder treeBuilder = new HtmlTreeBuilder();
         Parser parser = new Parser(treeBuilder);
         parser.errors = errorList;
-        return treeBuilder.parseFragment(fragmentHtml, context, baseUri, parser);
+        return treeBuilder.parseFragment(new StringReader(fragmentHtml), context, baseUri, parser);
     }
 
     /**
@@ -214,7 +219,7 @@ public class Parser {
      */
     public static List<Node> parseXmlFragment(String fragmentXml, String baseUri) {
         XmlTreeBuilder treeBuilder = new XmlTreeBuilder();
-        return treeBuilder.parseFragment(fragmentXml, null, baseUri, new Parser(treeBuilder));
+        return treeBuilder.parseFragment(new StringReader(fragmentXml), null, baseUri, new Parser(treeBuilder));
     }
 
     /**